requirements.txt
pyqt6
pyinstaller
keyboard
python-dateutil
pytest